Add emulation of PostgreSQL's search_path option
This will respect a :search_path Database option.  It takes the
schemas you pass in, and finds all tables/views in the schemas.
For any unqualified tables that are used in FROM/JOIN, it will
implicitly qualify them if it can recognize the table name.

Note that this is not perfect emulation.  This only affects
datasets, so it won't affect DDL queries, and won't affect
unqualified table references outside bare usage in FROM/JOIN,
such as passing a table identifier or name to a function.

For performance, this caches the tables in each schema on first
need, so it will not pick up changes to the schema that occur
after that point.

While here, fix tables/views to not require a JDBC specific
method.