Remove embedded newlines when using :empty_null=>:ruby (Fixes #4)
Don't rescue errors in the thread, and reraise them later, just
join the thread, which will raise the error.

DRY up the column handling by adding a write_col lambda, and
DRY up the hdfs up code using a variable.

Close the File object for the CSV file in the ensure block to
avoid a temporary leak.