Ensure we've pulled payload length from SASL
Seems like sometimes when the client would read the payload, the
length of the returned string was less than expected.  This would
cause the client to misread the length of the incoming payload.

So, so long as the server hasn't returned the full payload length,
we keep reading.  This seems to have fixed the error, hokey as
it may seem.