Refactor profile code to not use last_cursor when it may not be available
This introduces the following API for profiles, Dataset#profile for
telling the dataset to profile the query when it is executed, and
Database#profile_for retrieving the profiles.  You can use these
methods without arguments if you only care about a single profile.
However, if you care about multiple profiles, you can pass a symbol
to Dataset#profile and later pass the same symbol to
Database#profile_for to get that specific profile.  This is thread
safe as a mutex is used to protect access to the hash storing
profiles.

Note that this only allows you to profile SELECT queries, not
UPDATE, INSERT, or DELETE or arbitrary queries executed via
Database#run.  However, I assume for impala, this may be sufficient.
If not, it's easy to modify this to handle those cases.