Allow use of impala adapter when gssapi doesn't support the OS
gssapi supports only the most popular operating systems.  As you
only need the gssapi support if you are going to be using SASL,
it may be better to only load the sasl_transport code in that case.
However, this at least allows use of impala if gssapi is not
supported.