Use #query instead of #execute for fetching rows
We were only able to gather profile information for queries that
succeeded, but not for any that raised an exception.

We had been using ExecuteAndWait, which is nice in that we don't hear
back from the database until the query is done

However ExecuteAndWait does not return any query handle information
should the query fail

Calling Query returns immediately after the Query is started, giving
us a handle to the query, and thereby a handle to the profile.

But that requires us to monitor the state of the query and only attempt
to fetch results once it is done (or not fetch if there was an
exception).