Fix prepared statement deletes by returning cached SQL
This was broken because the delete_sql implementation was custom
and didn't return a cached value for a prepared statement.  As
prepared_args are frozen after the SQL is originally prepared
(in order to catch bugs at runtime such as this), this broke
with the refactoring of the prepared statement code in Sequel
4.42.0.

While here, make sure delete_sql will work with frozen
literal strings.