Update identifier mangling code
Identifier mangling is now optional in Sequel, though it is still
enabled by default.  This updates the shared impala adapter to
use the new way of handling identifier mangling.

This bumps the minimum supported Sequel version to 4.42.0. It is
possible to write the adapter such that it works on older versions,
but it basically requires having both the old and new code and
choosing which to execute based on the Sequel version.