Fix CTE usage with UNION/INTERSECT/EXCEPT
CTEs are not supported directly anymore, so they need to be hoisted.
Hoisting them fixed UNION, but showed a bug in INTERSECT emulation,
which is that INTERSECT emulation didn't handle NULL values since
6b249084895f457745b95a1c620143b323153e90 changed the emulation to
use an INNER JOIN.

This adds specific checks for NULL values on both sides of the JOIN
when emulating INTERSECT.  There could potentially be performance
issues with this, at least on databases where an index could be used
in the USING case but not in this case.  However, I don't think that
applies to Impala.  If so, we can make the behavior configurable via
an option.