Work with Sequel 5
This also should work with later versions of Sequel 4, but I'm not
sure how far back (possibly 4.42, possibly not).

There are also a couple of changes to allow working with frozen
literal strings.  While sequel-impala doesn't use them currently,
the adapters that ship with Sequel do.

This hasn't been extensively tested, and other changes may be
required for full compatibility.