Automatically run USE for new connections if :database is specified
The ruby impala driver doesn't take a database argument, so it
always connects to the default database.  This issues a USE
statement so the default database is set for each new connection,
so it works similar to other databases.

This probably won't help JDBC connections unless the :database
option is specified manually, since the JDBC connection strings
are not parsed by Sequel, but hopefully the JDBC drivers handle
the database part of the URL correctly.