Fix automatic use of USE to set database when using sharding
The previous approach didn't handle sharding.

This removes the definition in connect in the shared adapter as
it's not possible to handle it correctly for the jdbc:hive2 and
jdbc:impala adapters without overriding connect in each adapter.
As I'm not sure if those jdbc adapters respect the database
given in the URL, I'm not going to modify them yet.