Work around Impala bug when using complex dataset in create_table with :as and :sort_by
This creates a temporary unsorted table, so it cuts performance
by about half, but that is better than getting an AnalysisException
complaining about the use of sort.columns and specifying that
SORT BY should be used instead.  This error appears related to
Impala bug 5339, which was fixed before the release of 2.9.0.