Support TESTS_DIV_MOD environment variable when testing
This allows you to only test a subset of the tests.  The environment
variable should containing two integers separated by a space. The
first integer is the divisor, and the second is the remainder.
Tests are sequentially numbered (starting at 0), and if the
environment variable is set, the test is skipped unless the test
number divided by the divisor is equal to the remainder.

As an example TESTS_DIV_MOD="2 0" will run half the tests, and
TESTS_DIV_MOD="2 1" will run the other half, TESTS_DIV_MOD="4 0"
will run a quarter of the tests, and values of "4 1", "4 2", and
"4 3" will run the other three quarters, respectively.  This allows
for simple test splitting or parallelization (when run on different
databases).