Support :semi_join_first option when inner joining
This runs a LEFT SEMI JOIN between the two tables, which will remove
rows in the left hand side that are not matching in the right hand
side, use that as a subquery and then run an INNER JOIN on that subquery
and the original right hand side.

This will not work correctly if there are multiple tables being joined
and the join call using the :semi_join_first option uses an expression
with a qualified reference to a table that isn't the first table or
the table being joined to.  For example, the following will not work:

  DB[:a].join(:b, :a_id=>:id).
    join(:c, {:id=>Sequel[:b][:id]}, semi_join_first: true)

this is because the reference to b.id is not valid as b is in the
LEFT SEMI JOIN subquery in the left hand side, and that cannot be
referenced outside the subquery during the INNER JOIN.

That could theoretically be worked around by replacing all qualified
references in the join expression that do not match the first table
alias in the query or the table being joined with the the alias of
the subquery.